requirements.txt
Flask==0.11.1
gunicorn==19.7.1
prometheus-client
ibmcloudenv
livereload
watson-developer-cloud~=1.3.4
mysql-connector-python==8.0.12
flask_wtf
googlemaps
